test-requirements.txt
tox
ansible-lint
molecule
cryptography
jinja2
mock
pytest
pytest-ansible
pytest-mock
pytest-xdist
pytest-html
pyyaml
flake8
flake8-bugbear
flake8-colors
flake8-docstrings
flake8-import-order
pep8-naming
nodeenv
pygithub
requests
yamllint==1.15.0
bashate
pycodestyle
shyaml
pylint
voluptuous

